[#27] Download translations with a hierarchy
The reason for this change is that most Javascript I18n plugins expect
the data in the same hierarchy that it is stored on as the locale yaml
files.

Currently, we have to come up with a conversion to store and retrieve
our keys differently to match the key to the translation. With this
change we allow the flexibility of allowing the end user (CopyCopter
client or some other javascript client) to pass a optional argument
formart with a type of hierarchy to get the data in the format of:

  {
    "en": {
      "test": {
        "one": "expected one",
        "two": "expected two"
      }
    }
  }